+++ packages/jest-mock-apollo/src/index.ts
@@ -1,7 +1,7 @@
-import createGraphQLClientFactory from './factory';
+import configureClient from './graphqlClient';
 import MockApolloLink from './MockApolloLink';
 
 export {MockApolloLink};
 export * from './types';
 
-export default createGraphQLClientFactory;
+export default configureClient;
